Bug 1196077 - [email] Refactor message_list into some smaller custom elements r=asuth
