+++ apps/email/js/cards/lst/default_vscroll_data.js
@@ -1,0 +1,33 @@
+// Default data used for the VScroll component, when data is not loaded yet for
+// display in the virtual scroll listing. The u2583 character is a solid bar
+// character, so a string of them gives a "redacted" look to a line. Just a
+// temporary placeholder until the real data comes in.
+define({
+  'isPlaceholderData': true,
+  'id': 'INVALID',
+  'author': {
+    'name': '\u2583\u2583\u2583\u2583\u2583\u2583\u2583\u2583',
+    'address': '',
+    'contactId': null
+  },
+  'to': [
+    {
+      'name': ' ',
+      'address': ' ',
+      'contactId': null
+    }
+  ],
+  'cc': null,
+  'bcc': null,
+  'date': '0',
+  'hasAttachments': false,
+  'snippet': '\u2583\u2583\u2583\u2583\u2583\u2583\u2583\u2583' +
+             '\u2583\u2583\u2583\u2583\u2583\u2583\u2583\u2583' +
+             '\u2583\u2583\u2583\u2583\u2583\u2583\u2583\u2583',
+  'isRead': true,
+  'isStarred': false,
+  'sendStatus': {},
+  'subject': '\u2583\u2583\u2583\u2583\u2583\u2583\u2583\u2583' +
+             '\u2583\u2583\u2583\u2583\u2583\u2583\u2583\u2583' +
+             '\u2583\u2583\u2583\u2583\u2583\u2583\u2583\u2583'
+});
