+++ apps/email/js/cards/lst/edit_header.html
@@ -1,0 +1,12 @@
+<section role="region">
+  <header>
+    <a href="#"
+       data-event="click:domEvt"
+       data-domevt-name="editHeaderClose"
+       role="button"
+       data-l10n-id="close-button">
+      <span class="icon icon-close"></span>
+    </a>
+    <h1 data-prop="headerNode"></h1>
+  </header>
+</section>
