+++ apps/email/js/cards/lst/edit_toolbar.html
@@ -1,0 +1,31 @@
+<!-- Toolbar for multi-edit state -->
+<ul class="bb-tablist" role="toolbar">
+  <li role="presentation">
+    <button data-prop="deleteBtn"
+            data-event="click:domEvt"
+            data-domevt-name="onDeleteMessages"
+            class="icon msg-delete-btn"
+            data-l10n-id="message-delete-button"></button>
+  </li>
+  <li role="presentation">
+    <button data-prop="starBtn"
+            data-event="click:domEvt"
+            data-domevt-name="onStarMessages"
+            class="icon msg-star-btn"
+            data-l10n-id="message-star-button"></button>
+  </li>
+  <li role="presentation">
+    <button data-prop="readBtn"
+            data-event="click:domEvt"
+            data-domevt-name="onMarkMessagesRead"
+            class="icon msg-mark-read-btn"
+            data-l10n-id="message-mark-read-button"></button>
+  </li>
+  <li role="presentation">
+    <button data-prop="moveBtn"
+            data-event="click:domEvt"
+            data-domevt-name="onMoveMessages"
+            class="icon msg-move-btn"
+            data-l10n-id="message-move-button"></button>
+  </li>
+</ul>
