+++ apps/email/js/cards/lst/msg_vscroll.html
@@ -1,0 +1,22 @@
+<div data-prop="vScrollContainer" class="msg-vscroll-container"></div>
+<!-- maintain vertical space for the syncing/sync more div's
+     regardless of their displayed status so we don't scroll them
+     out of the way -->
+<div class="msg-messages-sync-container">
+  <p data-prop="syncingNode" class="msg-messages-syncing collapsed"
+     role="progressbar" data-l10n-id="messages-syncing-progressbar">
+    <span data-l10n-id="messages-syncing"></span>
+  </p>
+  <p data-prop="syncMoreNode"
+     data-event="click:onGetMoreMessages"
+     class="msg-messages-sync-more collapsed"
+     role="button">
+    <span data-l10n-id="messages-load-more"></span>
+  </p>
+</div>
+<div data-prop="messageEmptyContainer"
+     class="msg-list-empty-container collapsed">
+  <p data-prop="messageEmptyText"
+     class="msg-list-empty-message-text"
+     data-l10n-id="messages-folder-empty"></p>
+</div>
