+++ apps/email/js/cards/lst/peep_dom.js
@@ -1,0 +1,8 @@
+'use strict';
+define(function() {
+  return {
+    update: function(peep) {
+      peep.element.textContent = peep.name || peep.address;
+    }
+  };
+});
