+++ apps/email/js/cards/message_list.html
@@ -26,67 +26,13 @@
             class="msg-list-header-folder-name"></span>
       <span data-prop="folderUnread"
             class="msg-list-header-folder-unread collapsed"></span>
-      </h1>
-  </header>
-</section>
-<!-- Multi-edit state header -->
-<section data-prop="editHeader"
-         class="msg-listedit-header collapsed" role="region">
-  <header>
-    <a href="#" data-event="click:setEditModeDone"
-       class="msg-listedit-cancel-btn" role="button"
-       data-l10n-id="close-button">
-      <span class="icon icon-close"></span>
-    </a>
-    <h1 data-prop="headerNode" class="msg-listedit-header-label">
     </h1>
   </header>
 </section>
-<!-- Search header -->
-<section role="region" data-prop="searchHeader"
-         class="msg-search-header msg-search-only">
-  <form role="search" data-event="submit:onSearchSubmit">
-    <button data-event="click:onCancelSearch"
-            class="msg-search-cancel" type="submit"
-            data-l10n-id="message-search-cancel"></button>
-    <p>
-      <input data-prop="searchInput" data-event="input:onSearchTextChange"
-             type="text" required="required" class="msg-search-text"
-             autocorrect="off"
-             inputmode="verbatim"
-             x-inputmode="verbatim"
-             dir="auto"
-             data-l10n-id="message-search-input" />
-      <button type="reset" data-l10n-id="form-clear-input" data-event="click:onClearSearch"></button>
-    </p>
-  </form>
-  <!-- Search filter switcher -->
-  <header class="msg-search-controls-bar">
-    <ul role="tablist" class="bb-tablist filter" data-type="filter">
-      <li role="presentation" class="msg-search-from msg-search-filter"
-          data-filter="author">
-        <a data-l10n-id="message-search-from" role="tab"
-          aria-selected="false"></a></li>
-      <li role="presentation" class="msg-search-to msg-search-filter"
-          data-filter="recipients">
-        <a data-l10n-id="message-search-to" role="tab"
-          aria-selected="false"></a></li>
-      <li role="presentation" class="msg-search-subject msg-search-filter"
-           data-filter="subject">
-        <a data-l10n-id="message-search-subject" role="tab"
-          aria-selected="false"></a>
-      </li>
-      <li role="presentation" class="msg-search-body msg-search-filter"
-           data-filter="body">
-        <a data-l10n-id="message-search-body" role="tab"
-          aria-selected="false"></a></li>
-      <li role="presentation" class="msg-search-body msg-search-filter"
-          data-filter="all">
-        <a data-l10n-id="message-search-all" role="tab"
-          aria-selected="true"></a></li>
-    </ul>
-  </header>
-</section>
+
+<!-- Multi-edit state header -->
+<lst-edit-header data-prop="editHeader"
+                       data-event="editHeaderClose"></lst-edit-header>
 <!-- Scroll region -->
 <div data-prop="scrollContainer" class="msg-list-scrollouter">
   <!-- exists so we can force a minimum height -->
@@ -104,31 +50,16 @@
                data-l10n-id="message-search-input" />
       </p>
     </form>
-    <div data-prop="messagesContainer" class="msg-messages-container"
-         role="listbox" aria-multiselectable="true">
-    </div>
-    <!-- maintain vertical space for the syncing/sync more div's
-         regardless of their displayed status so we don't scroll them
-         out of the way -->
-    <div class="msg-messages-sync-container">
-      <p data-prop="syncingNode" class="msg-messages-syncing collapsed"
-         role="progressbar" data-l10n-id="messages-syncing-progressbar">
-        <span data-l10n-id="messages-syncing"></span>
-      </p>
-      <p data-prop="syncMoreNode"
-         data-event="click:onGetMoreMessages"
-         class="msg-messages-sync-more collapsed"
-         role="button">
-        <span data-l10n-id="messages-load-more"></span>
-      </p>
-    </div>
+    <lst-msg-vscroll data-prop="msgVScroll"
+                     data-event="messageClick:onClickMessage"
+                     data-empty-l10n-id="messages-folder-empty">
+    </lst-msg-vscroll>
   </div>
 </div>
+
 <!-- New email notification bar -->
 <div class="message-list-topbar"></div>
-<!-- Conveys background send, plus undo-able recent actions -->
-<div class="msg-activity-infobar hidden">
-</div>
+
 <!-- Toolbar for non-multi-edit state -->
 <ul data-prop="normalToolbar" class="bb-tablist msg-list-action-toolbar"
     role="toolbar">
@@ -151,35 +82,8 @@
   </li>
 </ul>
 
-<!-- Toolbar for multi-edit state -->
-<ul data-prop="editToolbar"
-    class="bb-tablist msg-listedit-action-toolbar collapsed"
-    role="toolbar">
-  <li role="presentation">
-    <button data-prop="deleteBtn" data-event="click:onDeleteMessages"
-            class="icon msg-delete-btn"
-            data-l10n-id="message-delete-button"></button>
-  </li>
-  <li role="presentation">
-    <button data-prop="starBtn" data-event="click:onStarMessages"
-            class="icon msg-star-btn"
-            data-l10n-id="message-star-button"></button>
-  </li>
-  <li role="presentation">
-    <button data-prop="readBtn" data-event="click:onMarkMessagesRead"
-            class="icon msg-mark-read-btn"
-            data-l10n-id="message-mark-read-button"></button>
-  </li>
-  <li role="presentation">
-    <button data-prop="moveBtn" data-event="click:onMoveMessages"
-            class="icon msg-move-btn"
-            data-l10n-id="message-move-button"></button>
-  </li>
-</ul>
+<lst-edit-toolbar data-prop="editToolbar"
+                  data-event="onDeleteMessages,onStarMessages,
+                              onMarkMessagesRead,onMoveMessages">
+</lst-edit-toolbar>
 
-<div data-prop="messageEmptyContainer"
-     class="msg-list-empty-container collapsed">
-  <p data-prop="messageEmptyText"
-     class="msg-list-empty-message-text"
-     data-l10n-id="messages-folder-empty"></p>
-</div>
