+++ apps/email/js/cards/mixins/data-event.js
@@ -6,6 +6,20 @@ define(function () {
     templateInsertedCallback: function () {
       slice.call(this.querySelectorAll('[data-event]'))
       .forEach(function (node) {
+        var parent = node;
+        // Make sure the node is not nested in another component.
+        while ((parent = parent.parentNode)) {
+          if (parent.nodeName.indexOf('-') !== -1) {
+            if (parent !== this) {
+              return;
+            }
+            break;
+          }
+        }
+        if (!parent) {
+          return;
+        }
+
         // Value is of type 'name:value,name:value',
         // with the :value part optional.
         node.dataset.event.split(',').forEach(function (pair) {
@@ -19,8 +33,8 @@ define(function () {
           method = parts[1].trim();
 
           if (typeof this[method] !== 'function') {
-            throw new Error('"' + method + '" is not a function, ' +
-                            'cannot bind with data-event');
+            throw new Error('"' + method +
+                            '" is not a function, cannot bind with data-event');
           }
 
           node.addEventListener(evtName, function(evt) {
