+++ apps/email/js/cards/settings_account_servers.js
@@ -9,7 +9,8 @@ var mozL10n = require('l10n!'),
     cards = require('cards');
 
 return [
-  require('./base')(require('template!./settings_account_servers.html')),
+  require('./base_card')
+         (require('template!./settings_account_servers.html')),
   {
     onArgs: function(args) {
       this.account = args.account;
