+++ apps/email/js/cards/setup_fix_gmail_twofactor.js
@@ -1,7 +1,8 @@
 'use strict';
 define(function(require) {
   return [
-    require('./base')(require('template!./setup_fix_gmail_twofactor.html')),
+    require('./base_card')
+           (require('template!./setup_fix_gmail_twofactor.html')),
     require('./setup_fix_mixin')
   ];
 });
