+++ apps/email/js/cards/setup_fix_oauth2.js
@@ -5,7 +5,7 @@ var oauthFetch = require('./oauth2/fetch'),
     cards = require('cards');
 
 return [
-  require('./base')(require('template!./setup_fix_oauth2.html')),
+  require('./base_card')(require('template!./setup_fix_oauth2.html')),
   {
     extraClasses: ['anim-fade', 'anim-overlay'],
