+++ apps/email/js/cards/setup_fix_password.js
@@ -4,7 +4,7 @@
 'use strict';
 define(function(require) {
   return [
-    require('./base')(require('template!./setup_fix_password.html')),
+    require('./base_card')(require('template!./setup_fix_password.html')),
     require('./setup_fix_mixin')
   ];
 });
