+++ apps/email/js/cards/setup_manual_config.js
@@ -12,7 +12,7 @@ var cards = require('cards'),
     plainSocketWarningNode = require('tmpl!./tng/plain_socket_warning.html');
 
 return [
-  require('./base')(require('template!./setup_manual_config.html')),
+  require('./base_card')(require('template!./setup_manual_config.html')),
   require('./setup_account_error_mixin'),
   {
     onArgs: function(args) {
